decodebin3: emit no-more-pads signal
The concept is: before exposing a new pad we check
if all the inputs are complete.
Then we know that the parsebins are not going to add more
streams: it can add and destroy pads, but they will be
associated to already created streams that are exposed.
So, if we have exposed a new pad, and the parsebin
is complete, we emit no-more-pads.